requirements.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
binaryornot==0.4.4
brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cookiecutter==2.6.0
crispy-tailwind==1.0.3
cryptography==45.0.3
cssselect2==0.8.0
dj-database-url==3.0.0
django==5.2.1
django-axes==8.0.0
django-browser-reload==1.18.0
django-crispy-forms==2.4
django-csp==4.0
django-formtools==2.5.1
django-material==1.12.0
django-pwa==2.0.1
django-tailwind==4.0.1
django-widget-tweaks==1.5.0
django-xhtml2pdf==0.0.4
fonttools==4.58.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
jinja2==3.1.6
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pygments==2.19.1
pyhanko==0.29.0
pyhanko-certvalidator==0.27.0
pypdf==5.5.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pyyaml==6.0.2
reportlab==4.4.1
requests==2.32.3
rich==14.0.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
types-python-dateutil==2.9.0.20250516
typing-extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
